Fix issue with $PATH not working for fish
Since /usr/bin/env relies on $PATH to resolve the shell that is to be executed, this breaks when the
shell in question (say /opt/homebrew/bin/fish) isn't in the default path (like /usr/bin).
This fixes this by passing PATH=<current path> to the newly formed sessions.

Secondly, unsetting the history file in fish works differently than in bash/zsh, for which this
commit provides an exception.

This should fix #146